# specs/openapi/dist/api-bundle.yaml
openapi: 3.1.0
info:
  title: Virtual Companion API
  version: 0.1.0
  description: Contract-first HTTP API (ADR-0002). This document is the single source of truth; Java interfaces and TypeScript
    clients are generated from it and must not be hand-edited or used to back-drive the contract. The ErrorEnvelope shape
    is uniform across every error response; NOT_FOUND_OR_FORBIDDEN hides resource existence so a missing resource and a forbidden
    one are indistinguishable to the caller.
security:
- bearerAuth: []
paths:
...
  /api/v1/age/state:
    get:
      operationId: getAgeState
      summary: The caller's effective age state (FR-AUTH-002)
      description: The effective adult-verification state of the caller (age-states catalog), AGE_UNKNOWN when never verified.
        The stored record carries only the result, the age band, the verification time and the provider reference — never
        an identity document. Requires a server-verified authentication context.
      responses:
        '200':
          description: The effective age state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgeStateResponse'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/age/verification:
    post:
      operationId: verifyAge
      summary: Run the simulated adult verification (FR-AUTH-002, Alpha port)
      description: Walks the age-states catalog transition table from the caller's current state to ADULT_VERIFIED through
        the vendor-agnostic AgeVerificationPort (Technical Alpha uses the deterministic simulated verifier; a real provider
        replaces the port implementation, never the callers). Every step is appended to the result history with the provider
        reference and verification time; no identity document is stored. A current state that the catalog forbids from reaching
        ADULT_VERIFIED (minor, appeal-pending, suspended) maps to INVALID_REQUEST (fail closed). Requires a server-verified
        authentication context.
      responses:
        '200':
          description: The new effective age state (ADULT_VERIFIED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgeStateResponse'
        '400':
          description: The current age state cannot complete adult verification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/age/appeal:
    post:
      operationId: submitAgeAppeal
      summary: Submit an age-verification appeal (FR-AUTH-002)
      description: Appends an age appeal for the caller (vc.submit_age_appeal) and flips the effective age state to AGE_APPEAL_PENDING
        in the same transaction. Allowed only from a state the age-states catalog can reach AGE_APPEAL_PENDING from (ADULT_VERIFICATION_REQUIRED
        / MINOR_SUSPECTED); anything else maps to INVALID_REQUEST (fail closed). Resolution stays a human review action —
        this endpoint only records the appeal, it never rewrites a verification result. Requires a server-verified authentication
        context.
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgeAppealRequest'
      responses:
        '200':
          description: The appended appeal record (status SUBMITTED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgeAppealRecord'
        '400':
          description: The current age state cannot submit an appeal, or an invalid body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/age/appeals:
    get:
      operationId: listAgeAppeals
      summary: The caller's age appeals, newest first (FR-AUTH-002)
      description: Keyset-paginated age appeals owned by the caller (vc.list_age_appeals). The after cursor is the last appeal
        id seen. Appeals are append-only intake records; resolution status changes only through human review. Requires a server-verified
        authentication context.
      security:
      - bearerAuth: []
      parameters:
      - name: after
        in: query
        required: false
        schema:
          type: string
        description: Opaque cursor (the last appeal id seen); omit to start from the beginning.
      - name: limit
        in: query
        required: false
        schema:
          type: integer
        description: Page size, clamped to a safe band by the server.
      responses:
        '200':
          description: A page of the caller's appeals, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgeAppealRecord'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorEnvelope:
      type: object
      description: Unified error shape. NOT_FOUND_OR_FORBIDDEN never discloses whether a resource exists; every error response
        uses this envelope.
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Stable, non-sensitive human-readable explanation
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (never sensitive data)
    ErrorCode:
      type: string
      description: Stable error code catalog (specs/catalog/error-codes.yaml)
      enum:
      - AUTHENTICATION_REQUIRED
      - ACCESS_DENIED
      - NOT_FOUND_OR_FORBIDDEN
      - AGE_VERIFICATION_REQUIRED
      - MODEL_NO_ELIGIBLE_DEPLOYMENT
      - MODEL_ATTEMPT_TIMEOUT
      - MODEL_ALL_UNAVAILABLE
      - GENERATION_CANCELLED
      - GENERATION_OUTPUT_BLOCKED
      - STREAM_GAP_EXPIRED
      - STREAM_RESET_REQUIRED
      - MEMORY_CONFIRMATION_REQUIRED
      - MEMORY_FORBIDDEN_CONTENT
      - CATALOG_DRIFT_DETECTED
      - BETA_OPERATIONS_NOT_READY
      - INVALID_REQUEST
      - AUTH_RATE_LIMITED
...
    AgeStateResponse:
      type: object
      description: The caller's effective adult-verification state (FR-AUTH-002). Only the result, the age band, the verification
        time and the provider reference are stored — never an identity document.
      required:
      - ageState
      properties:
        ageState:
          type: string
          description: age-states catalog code (AGE_UNKNOWN when never verified)
          enum:
          - AGE_UNKNOWN
          - ADULT_SELF_DECLARED
          - ADULT_VERIFICATION_REQUIRED
          - ADULT_VERIFIED
          - MINOR_SUSPECTED
          - MINOR_VERIFIED
          - AGE_APPEAL_PENDING
          - AGE_REVERIFY_REQUIRED
          - AGE_ACCESS_SUSPENDED
        providerRef:
          type: string
          description: Verification provider reference (alpha-simulated in Technical Alpha)
        verifiedAt:
          type: string
          format: date-time
          description: Verification time (RFC 3339), present once verified
    AgeAppealRequest:
      type: object
      description: Age-appeal submission body (FR-AUTH-002). The reason is the caller's own appeal statement; it is stored
        trimmed and never rewrites a verification result.
      required:
      - reason
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
          description: Why the caller believes the age verdict is wrong (trimmed server-side).
    AgeAppealRecord:
      type: object
      description: An age-appeal intake record (FR-AUTH-002). SUBMITTED until a human review resolves it; resolution fields
        are present only once resolved.
      required:
      - id
      - reason
      - status
      - createdAt
      properties:
        id:
          type: string
          description: The appeal id (opaque cursor for listAgeAppeals).
        reason:
          type: string
          description: The caller's appeal statement as submitted.
        status:
          type: string
          enum:
          - SUBMITTED
          - RESOLVED
          description: Intake lifecycle status; resolution is a human review action.
        resolutionNote:
          type: string
          description: Present once a human review resolved the appeal.
        createdAt:
          type: string
          format: date-time
          description: Submission time (RFC 3339).
        resolvedAt:
          type: string
          format: date-time
          description: Resolution time (RFC 3339), present once resolved.
...